SCRUM-210 feat: update Payments component with pagination and refactor styles
