+++ package.json
@@ -28,7 +28,7 @@
     "@fontsource-variable/inter": "^5.2.6",
     "@fontsource-variable/roboto": "^5.2.6",
     "@hookform/resolvers": "^5.1.1",
-    "@ictroot/ui-kit": "2.0.0-beta.1",
+    "@ictroot/ui-kit": "2.0.0-beta.3",
     "@radix-ui/react-alert-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
     "@radix-ui/react-slider": "^1.3.6",
