+++ src/shared/composites/Table/Table.module.scss
@@ -3,6 +3,7 @@
   border-collapse: collapse;
   font-size: var(--font-size-s);
   border: 1px solid var(--color-dark-500);
+  border-top: none;
 }
 
 .tableHeader {
